- Upgrade to scons 3.0.0 - Fix build error witn vs 2017 - Fix build error with Python 3.0+

+++ src/private/core/gfx/misc/asciifont.cpp
@@ -56,7 +56,7 @@ public:
 
         const BitmapCharDesc * bcd = gBitmapChars8x13[ch];
 
-        std::fill( mImage, mImage+sizeof(mImage), 0 );
+        std::fill( mImage, mImage+sizeof(mImage), (uint8)0 );
 
         for( size_t y = 0; y < bcd->height; ++y )
         {
