+++ src/private/core/gfx/misc/bitmapFont.cpp
@@ -295,7 +295,7 @@ GN::gfx::BitmapFont::createSlot( wchar_t ch )
     // copy font image into RGBA
     GN_ASSERT( fbm.width <= slot.w && fbm.height <= slot.h );
     DynaArray<uint8> tmpbuf( slot.w * slot.h * 4 );
-    std::fill( tmpbuf.begin(), tmpbuf.end(), 0 );
+    std::fill( tmpbuf.begin(), tmpbuf.end(), (uint8)0 );
     for( size_t y = 0; y < fbm.height; ++y )
     {
         for( size_t x = 0; x < fbm.width; ++x )
